# Include Mutex for Hubs property