@ethervoid 2020-02-18
@agnivade this PR is open to discussion because I'm not pretty sure it is needed maybe is only happening in tests, but if the performance is not affected could be a good point to include and fix some future problems in the server and current ones in the tests.

Another question [here](https://github.com/mattermost/mattermost-server/pull/13884#pullrequestreview-359379224) you asks for benchmark, where can I get that benchmark?