@ethervoid 2020-02-19
@agnivade 

> I'm trying to understand how this race condition may be a fix for the test? Can you explain?

I've been unable to reproduce the test failing. I've tried multiple ways and I've read the code but after multiple and intermittent days working on this, it didn't fail. 

The only problem I've found is this race condition in the tests and given is a race condition it could have interfered in some weird way but I can't connect them in any way.

In the end, this PR is to fix a possible race condition rather than fix the flaky test

> Re: the benchmarks, they are not there, you will have to make them yourselves.

Yes, I'm aware but my question is more focused on what kind of benchmarks do you do and what tooling do you use.