@agnivade 2020-02-19
> In the end, this PR is to fix a possible race condition rather than fix the flaky test

Ok, gotcha.

> Yes, I'm aware but my question is more focused on what kind of benchmarks do you do and what tooling do you use.

Just some simple benchmarks trying to add/remove the hubs from multiple goroutines should be sufficient. Have a version with single goroutine to compare with the current code.