@ethervoid 2020-02-19
> Just some simple benchmarks trying to add/remove the hubs from multiple goroutines should be sufficient. Have a version with single goroutine to compare with the current code.

Oh, ok. Thank you

I'll make some benchmarks to make sure there are not regressions given that adding synchronization allows the scheduler to switch the 
current goroutine for another one or in the worst case having the current goroutine in waiting state.

If there aren't regressions we could merge this code to fix the test's race condition and try to avoid future problems. What do you think?