@ethervoid 2020-02-20
@agnivade 

> I'm fine with merging this. But then let's make a separate ticket, file it under Race conditions epic and link this to that. And let the flaky test remain as a separate ticket.

Done. You can see the new ticket link in the PR description