@streamer45 2020-02-20
@ethervoid I am wondering if this does also fix the hub stop data race?