@streamer45 2020-02-20
> @ethervoid I am wondering if this does also fix the hub stop data race?

Actually I don't think we had a ticket to track that. I've made one https://mattermost.atlassian.net/browse/MM-22656