@ethervoid 2020-02-20
@streamer45 Yes, this PR fixes that data race too :)