@ethervoid 2020-02-20
The comparison using `benchstat`:
```
name                    old time/op    new time/op    delta
SingleGoRoutineHubs-16     311ns ±19%     483ns ±17%  +55.28%  (p=0.000 n=10+8)
TenGoRoutineHubs-16       2.94µs ±41%    5.83µs ±84%  +98.17%  (p=0.000 n=9+10)

name                    old alloc/op   new alloc/op   delta
SingleGoRoutineHubs-16     98.0B ± 0%    184.2B ±12%  +88.01%  (p=0.000 n=9+8)
TenGoRoutineHubs-16         981B ± 0%     1891B ±20%  +92.79%  (p=0.000 n=9+8)

name                    old allocs/op  new allocs/op  delta
SingleGoRoutineHubs-16      1.00 ± 0%      1.00 ± 0%     ~     (all equal)
TenGoRoutineHubs-16         10.0 ± 0%      19.0 ± 0%  +90.00%  (p=0.000 n=9+8)
```