@ethervoid 2020-02-21
Adding the new method el benchmark is a bit worse:
```
name                    old time/op    new time/op    delta
SingleGoRoutineHubs-16     311ns ±19%     623ns ±73%  +100.51%  (p=0.000 n=10+10)
TenGoRoutineHubs-16       2.94µs ±41%    6.21µs ±80%  +111.10%  (p=0.000 n=9+10)

name                    old alloc/op   new alloc/op   delta
SingleGoRoutineHubs-16     98.0B ± 0%    186.1B ±17%   +89.92%  (p=0.000 n=9+8)
TenGoRoutineHubs-16         981B ± 0%     2136B ±53%  +117.85%  (p=0.000 n=9+9)

name                    old allocs/op  new allocs/op  delta
SingleGoRoutineHubs-16      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
TenGoRoutineHubs-16         10.0 ± 0%      19.0 ± 0%   +90.00%  (p=0.000 n=9+7)
 ``