@ethervoid 2020-02-21
Here is the benchmark for CPU 2,4,8,16
```
name                    old time/op    new time/op    delta
HubsSingleGoRoutine-2      421ns ±48%     659ns ±16%   +56.35%  (p=0.000 n=10+9)
HubsSingleGoRoutine-4     356ns ±128%     628ns ±97%   +76.56%  (p=0.028 n=10+10)
HubsSingleGoRoutine-8      244ns ±11%     584ns ±85%  +139.40%  (p=0.000 n=9+10)
HubsSingleGoRoutine-16     277ns ± 4%     592ns ±83%  +113.63%  (p=0.000 n=8+20)
HubsTenGoRoutine-2        3.64µs ±30%    6.05µs ±20%   +66.16%  (p=0.000 n=9+9)
HubsTenGoRoutine-4        2.06µs ± 6%    4.40µs ± 9%  +113.52%  (p=0.000 n=9+8)
HubsTenGoRoutine-8        2.41µs ± 7%    5.54µs ±96%  +129.65%  (p=0.000 n=9+10)
HubsTenGoRoutine-16       2.82µs ±19%    6.23µs ±79%  +121.30%  (p=0.000 n=9+20)

name                    old alloc/op   new alloc/op   delta
HubsSingleGoRoutine-2      98.7B ± 1%    144.5B ± 4%   +46.45%  (p=0.000 n=9+8)
HubsSingleGoRoutine-4       118B ±87%      218B ±64%   +85.27%  (p=0.001 n=9+9)
HubsSingleGoRoutine-8      98.0B ± 0%    191.5B ±10%   +95.41%  (p=0.000 n=8+8)
HubsSingleGoRoutine-16     98.0B ± 0%    187.9B ±16%   +91.71%  (p=0.000 n=9+16)
HubsTenGoRoutine-2          988B ± 1%     1466B ± 5%   +48.39%  (p=0.000 n=9+9)
HubsTenGoRoutine-4          977B ± 0%     1935B ± 0%   +97.96%  (p=0.000 n=8+7)
HubsTenGoRoutine-8          979B ± 0%     1933B ± 1%   +97.31%  (p=0.000 n=9+7)
HubsTenGoRoutine-16         981B ± 0%     1962B ±22%   +99.98%  (p=0.000 n=9+16)

name                    old allocs/op  new allocs/op  delta
HubsSingleGoRoutine-2       1.00 ± 0%      1.00 ± 0%      ~     (all equal)
HubsSingleGoRoutine-4       1.00 ± 0%      1.30 ±54%      ~     (p=0.211 n=10+10)
HubsSingleGoRoutine-8       1.00 ± 0%      1.00 ± 0%      ~     (all equal)
HubsSingleGoRoutine-16      1.00 ± 0%      1.00 ± 0%      ~     (all equal)
HubsTenGoRoutine-2          10.0 ± 0%      14.3 ± 5%   +43.33%  (p=0.000 n=9+9)
HubsTenGoRoutine-4          10.0 ± 0%      19.0 ± 0%   +90.00%  (p=0.000 n=9+9)
HubsTenGoRoutine-8          10.0 ± 0%      19.0 ± 0%   +90.00%  (p=0.000 n=9+8)
HubsTenGoRoutine-16         10.0 ± 0%      19.0 ± 0%   +90.00%  (p=0.000 n=9+14)
```